use pagination helper function in teamsconnection

+++ frontend/src/graphql/queries/newsConnection.ts
@@ -15,7 +15,7 @@ export interface NewsConnectionInput extends PaginationInput {
 export const NEWS_CONNECTION_QUERY = gql`
   query NewsConnection(
     $teamId: Int!
-    $first: Float
+    $first: Int
     $before: String
     $after: String
   ) {
